+++ negotiation/src/main/java/it/eng/negotiation/service/AgreementAPIService.java
@@ -44,7 +44,7 @@ public class AgreementAPIService {
         // TODO add additional checks like contract dates
         //		LocalDateTime agreementStartDate = LocalDateTime.parse(agreement.getTimestamp(), FORMATTER);
         //		agreementStartDate.isBefore(LocalDateTime.now());
-
+        log.debug("Found Agreement with DSP id {}", agreementId);
         PolicyDecision policyDecision = policyEnforcementPoint.enforcePolicy(agreement, "enforceAgreement");
 
         if (policyDecision.isAllowed()) {